reviewed RangeHolder.getRange() use
It is called many times during sort.
Luckily both RangeSortEntry and RangeForOverlapCheck stores
a copy of the range, does not need to ask the document.

RangeHolder.getRange().getText() is still called a few (up to about 3?) times.
We might gain some speed if it is very costly. But the time used
is probably dominated by the range comparisons.